refactor: extract OnboardingSteps and ImplementationCodeBlock into own files
